align with PR feedback (Will)
